backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-auth==2.29.0
google-cloud-aiplatform==1.44.0
python-dotenv==1.0.1
websockets==12.0
